move options constants to constants directory
