+++ frontend/src/constants/searchFilterOptions.ts
